+++ example/example/spiders/books_02_2.py
@@ -11,12 +11,12 @@ has problems now, it is used in the latter examples, because as an API
 it is better than defining callback explicitly.
 """
 import scrapy
-from web_poet import ItemWebPage
+from web_poet import WebPage
 
 from scrapy_poet import callback_for
 
 
-class BookPage(ItemWebPage):
+class BookPage(WebPage):
     def to_item(self):
         return {
             "url": self.url,