+++ example/example/spiders/books_04_overrides_03.py
@@ -11,7 +11,7 @@ difference is that this example is using the ``@handle_urls`` decorator to
 store the rules in web-poet's registry.
 """
 import scrapy
-from web_poet import ItemWebPage, WebPage, default_registry, handle_urls
+from web_poet import WebPage, default_registry, handle_urls
 
 from scrapy_poet import callback_for
 
@@ -21,12 +21,11 @@ class BookListPage(WebPage):
         return []
 
 
-class BookPage(ItemWebPage):
-    def to_item(self):
-        return None
+class BookPage(WebPage):
+    pass
 
 
-@handle_urls("toscrape.com", overrides=BookListPage)
+@handle_urls("toscrape.com", instead_of=BookListPage)
 class BTSBookListPage(BookListPage):
     """Logic to extract listings from pages like https://books.toscrape.com"""
 
@@ -34,7 +33,7 @@ class BTSBookListPage(BookListPage):
         return self.css(".image_container a::attr(href)").getall()
 
 
-@handle_urls("toscrape.com", overrides=BookPage)
+@handle_urls("toscrape.com", instead_of=BookPage)
 class BTSBookPage(BookPage):
     """Logic to extract book info from pages like https://books.toscrape.com/catalogue/soumission_998/index.html"""
 
@@ -45,7 +44,7 @@ class BTSBookPage(BookPage):
         }
 
 
-@handle_urls("bookpage.com", overrides=BookListPage)
+@handle_urls("bookpage.com", instead_of=BookListPage)
 class BPBookListPage(BookListPage):
     """Logic to extract listings from pages like https://bookpage.com/reviews"""
 
@@ -53,7 +52,7 @@ class BPBookListPage(BookListPage):
         return self.css("article.post h4 a::attr(href)").getall()
 
 
-@handle_urls("bookpage.com", overrides=BookPage)
+@handle_urls("bookpage.com", instead_of=BookPage)
 class BPBookPage(BookPage):
     """Logic to extract from pages like https://bookpage.com/reviews/25879-laird-hunt-zorrie-fiction"""
 
@@ -68,7 +67,7 @@ class BooksSpider(scrapy.Spider):
     name = "books_04_overrides_03"
     start_urls = ["http://books.toscrape.com/", "https://bookpage.com/reviews"]
     # Configuring different page objects pages for different domains
-    custom_settings = {"SCRAPY_POET_OVERRIDES": default_registry.get_overrides()}
+    custom_settings = {"SCRAPY_POET_OVERRIDES": default_registry.get_rules()}
 
     def parse(self, response, page: BookListPage):
         yield from response.follow_all(page.book_urls(), callback_for(BookPage))