+++ scrapy_poet/overrides.py
@@ -1,17 +1,18 @@
 import logging
 from abc import ABC, abstractmethod
 from collections import defaultdict
-from typing import Callable, Dict, Iterable, List, Mapping, Optional, Tuple, Union
+from typing import Callable, Dict, Iterable, List, Mapping, Optional, Tuple, Type, Union
 
 from scrapy import Request
 from scrapy.crawler import Crawler
 from url_matcher import Patterns, URLMatcher
-from web_poet.overrides import OverrideRule
+from web_poet import ItemPage
+from web_poet.rules import ApplyRule
 
 logger = logging.getLogger(__name__)
 
-RuleAsTuple = Union[Tuple[str, Callable, Callable], List]
-RuleFromUser = Union[RuleAsTuple, OverrideRule]
+RuleAsTuple = Union[Tuple[str, Type[ItemPage], Type[ItemPage]], List]
+RuleFromUser = Union[RuleAsTuple, ApplyRule]
 
 
 class OverridesRegistryBase(ABC):
@@ -29,7 +30,7 @@ class OverridesRegistry(OverridesRegistryBase):
     """
     Overrides registry that reads the overrides from the ``SCRAPY_POET_OVERRIDES``
     in the spider settings. It is a list and each rule can be a tuple or an
-    instance of the class :py:class:`web_poet.overrides.OverrideRule`.
+    instance of the class :py:class:`web_poet.rules.ApplyRule`.
 
     If a tuple is provided:
 
@@ -45,7 +46,7 @@ class OverridesRegistry(OverridesRegistryBase):
     .. code-block:: python
 
         from url_matcher import Patterns
-        from scrapy_poet.overrides import OverrideRule
+        from web_poet.rules import ApplyRule
 
 
         SCRAPY_POET_OVERRIDES = [
@@ -53,7 +54,7 @@ class OverridesRegistry(OverridesRegistryBase):
             ("books.toscrape.com", ISBNBookPage, BookPage),
 
             # Option 2
-            OverrideRule(
+            ApplyRule(
                 for_patterns=Patterns(["books.toscrape.com"]),
                 use=MyBookListPage,
                 instead_of=BookListPage,
@@ -63,12 +64,12 @@ class OverridesRegistry(OverridesRegistryBase):
     .. _web-poet: https://web-poet.readthedocs.io
 
     Now, if you've used web-poet_'s built-in functionality to directly create
-    the :py:class:`web_poet.overrides.OverrideRule` in the Page Object via the
+    the :py:class:`web_poet.rules.ApplyRule` in the Page Object via the
     :py:func:`web_poet.handle_urls` annotation, you can quickly import them via
     the following code below. It finds all the rules annotated using web-poet_'s
     :py:func:`web_poet.handle_urls` as a decorator that were registered into
     ``web_poet.default_registry`` (an instance of
-    :py:class:`web_poet.overrides.PageObjectRegistry`).
+    :py:class:`web_poet.rules.RulesRegistry`).
 
     .. code-block:: python
 
@@ -78,9 +79,9 @@ class OverridesRegistry(OverridesRegistryBase):
         # import rules from other packages. Otherwise, it can be omitted.
         # More info about this caveat on web-poet docs.
         consume_modules("external_package_A.po", "another_ext_package.lib")
-        SCRAPY_POET_OVERRIDES = default_registry.get_overrides()
+        SCRAPY_POET_OVERRIDES = default_registry.get_rules()
 
-    Make sure to call :py:func:`web_poet.overrides.consume_modules` beforehand.
+    Make sure to call :py:func:`web_poet.rules.consume_modules` beforehand.
     More info on this at web-poet_.
     """
 
@@ -89,26 +90,27 @@ class OverridesRegistry(OverridesRegistryBase):
         return cls(crawler.settings.getlist("SCRAPY_POET_OVERRIDES", []))
 
     def __init__(self, rules: Optional[Iterable[RuleFromUser]] = None) -> None:
-        self.rules: List[OverrideRule] = []
-        self.matcher: Dict[Callable, URLMatcher] = defaultdict(URLMatcher)
+        self.rules: List[ApplyRule] = []
+        self.matcher: Dict[Type[ItemPage], URLMatcher] = defaultdict(URLMatcher)
         for rule in rules or []:
             self.add_rule(rule)
-        logger.debug(f"List of parsed OverrideRules:\n{self.rules}")
+        logger.debug(f"List of parsed ApplyRules:\n{self.rules}")
 
     def add_rule(self, rule: RuleFromUser) -> None:
         if isinstance(rule, (tuple, list)):
             if len(rule) != 3:
                 raise ValueError(
-                    f"Invalid overrides rule: {rule}. Rules as tuples must have "
+                    f"Invalid rule: {rule}. Rules as tuples must have "
                     f"3 elements: (1) the pattern, (2) the PO class used as a "
                     f"replacement and (3) the PO class to be replaced."
                 )
             pattern, use, instead_of = rule
-            rule = OverrideRule(
+            rule = ApplyRule(
                 for_patterns=Patterns([pattern]), use=use, instead_of=instead_of
             )
         self.rules.append(rule)
-        self.matcher[rule.instead_of].add_or_update(
+        # FIXME: This key will change with the new rule.to_return
+        self.matcher[rule.instead_of].add_or_update(  # type: ignore
             len(self.rules) - 1, rule.for_patterns
         )