+++ tests/po_lib/__init__.py
@@ -4,7 +4,7 @@ This package is just for overrides testing purposes.
 import socket
 
 from url_matcher.util import get_domain
-from web_poet import ItemWebPage, handle_urls
+from web_poet import WebPage, handle_urls
 
 from tests.mockserver import get_ephemeral_port
 
@@ -13,12 +13,12 @@ DOMAIN = get_domain(socket.gethostbyname(socket.gethostname()))
 PORT = get_ephemeral_port()
 
 
-class POOverriden(ItemWebPage):
+class POOverriden(WebPage):
     def to_item(self):
-        return {"msg": "PO that will be replace"}
+        return {"msg": "PO that will be replaced"}
 
 
-@handle_urls(f"{DOMAIN}:{PORT}", overrides=POOverriden)
-class POIntegration(ItemWebPage):
+@handle_urls(f"{DOMAIN}:{PORT}", instead_of=POOverriden)
+class POIntegration(WebPage):
     def to_item(self):
         return {"msg": "PO replacement"}